Move functions related to CoM to new jaxsim.api.com module
