+++ src/jaxsim/api/__init__.py
@@ -1,3 +1,3 @@
 from . import common  # isort:skip
 from . import model, data  # isort:skip
-from . import contact, joint, kin_dyn_parameters, link, ode, ode_data, references
+from . import com, contact, joint, kin_dyn_parameters, link, ode, ode_data, references
